throttling structure improvement
